# src/routes/swagger.json
{
	"components": {
...
		"schemas": {
...
			"UpdatePointByTeacherRequest": {
				"properties": {
					"classId": {
						"type": "string"
					},
					"MSSV": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"assignmentId": {
						"type": "string"
					},
					"point": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"classId",
					"MSSV",
					"fullName",
					"assignmentId",
					"point"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"oauth2": {
				"type": "apiKey",
				"name": "authorization",
				"in": "header",
				"authorizationUrl": "http://swagger.io/api/oauth/dialog",
				"flow": "implicit",
				"scopes": {
					"write:pets": "modify things",
					"read:pets": "read things"
				}
			}
		}
	},
	"info": {
		"title": "Data Entry API",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/files/upload-list-student": {
			"post": {
				"operationId": "UploadFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Files"
				],
				"security": [
					{
						"oauth2": []
					}
				],
				"parameters": []
			}
		},
		"/files/list-student/{classId}": {
			"get": {
				"operationId": "ListStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Files"
				],
				"security": [
					{
						"oauth2": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "classId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/point-assignment/update-by-teacher": {
			"post": {
				"operationId": "UpdatePointByTeacher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Point Assingments"
				],
				"security": [
					{
						"oauth2": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePointByTeacherRequest"
							}
						}
					}
				}
			}
		},
		"/point-assignment/show-full-point/{classId}": {
			"get": {
				"operationId": "ShowFullPointInClass",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Point Assingments"
				],
				"security": [
					{
						"oauth2": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "classId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/point-assignment/show-point-assignment/{classId}/{assignmentId}": {
			"get": {
				"operationId": "ShoưPointAssignmentInClass",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Point Assingments"
				],
				"security": [
					{
						"oauth2": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "classId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "assignmentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/point-assignment/upload-file-point-by-teacher": {
			"post": {
				"operationId": "UploadPointByTeacher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Point Assingments"
				],
				"security": [
					{
						"oauth2": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}